Added methods to opi_helper to structure packages
Refactored many code segments in the biofeedback module to make it more
readable.
Also added two methods (buildDC01SDC01 and buildDC10) which build the
structures of status report from the controller and the data packets
from the TrueSense sensor. This makes the data handling much easier and
these are also the preparation for the live EDF writer.